Changed filename "ConfigureServices.cs" to "DependencyInjection.cs"
