--- src/SeeSharp.Application/ConfigureServices.cs
@@ -1,26 +1,0 @@
-﻿using System.Reflection;
-using MediatR;
-using Microsoft.Extensions.DependencyInjection;
-using SeeSharp.Application.Common.Behaviors;
-using FluentValidation;
-
-namespace SeeSharp.Infrastructure;
-
-public static class DependencyInjection
-{
-    public static IServiceCollection AddApplicationServices(this IServiceCollection services)
-    {
-        services.AddAutoMapper(Assembly.GetExecutingAssembly());
-        services.AddValidatorsFromAssembly(Assembly.GetExecutingAssembly());
-
-        services.AddMediatR(config =>
-        {
-            config.RegisterServicesFromAssembly(Assembly.GetExecutingAssembly());
-            config.AddBehavior(typeof(IPipelineBehavior<,>), typeof(LoggingPipelineBehavior<,>));
-            config.AddBehavior(typeof(IPipelineBehavior<,>), typeof(ValidationPipelineBehavior<,>));
-        });
-
-        return services;
-    }
-}
-
