+++ src/SeeSharp.GraphQL/Query/Query.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace SeeSharp.GraphQL.Query;
-internal class Query
+namespace SeeSharp.GraphQL;
+public class Query
 {
 }
