+++ src/SeeSharp.GraphQL/SeeSharp.GraphQL.csproj
@@ -6,4 +6,8 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="HotChocolate.AspNetCore" Version="13.9.0" />
+  </ItemGroup>
+
 </Project>
